+++ src/index.ts
@@ -18,15 +18,16 @@ export * from './DMA/DMA.js';
 export * from './DX/DX.js';
 export * from './EMA/EMA.js';
 export * from './error/index.js';
+export * from './IQR/IQR.js';
 export * from './LINREG/LinearRegression.js';
 export * from './MACD/MACD.js';
 export * from './MAD/MAD.js';
 export * from './MOM/MOM.js';
 export * from './OBV/OBV.js';
+export * from './PSAR/PSAR.js';
 export * from './RMA/RMA.js';
 export * from './ROC/ROC.js';
 export * from './RSI/RSI.js';
-export * from './PSAR/PSAR.js';
 export * from './SMA/SMA.js';
 export * from './STOCH/StochasticOscillator.js';
 export * from './STOCH/StochasticRSI.js';
